Create an image preloaded with datasets required for HPO lookup.  Relace yaml2ped with json2ped.
Store the data in the git repo.
