+++ docker/wgs_tertiary/build.env
@@ -1,0 +1,10 @@
+# Image revision
+IMAGE_BUILD=1
+
+# Software version
+ZENODO_RECORD=8415406
+WORKFLOW_VERSION=1.0.2
+
+# Image info
+IMAGE_NAME=wgs_tertiary
+IMAGE_TAG=${ZENODO_RECORD}_${WORKFLOW_VERSION}_build${IMAGE_BUILD}
