+++ docker/wgs_tertiary/data/hpo/README
@@ -1,0 +1,3 @@
+Last generated on 20230428.
+
+$ bash ./buildHpo
